+++ src/test/scala/edu/rit/cs/dataunitylab/jsonoid/discovery/transformers/DefinitionTransformerSpec.scala
@@ -1,4 +1,5 @@
 package edu.rit.cs.dataunitylab.jsonoid.discovery
+package transformers
 
 import org.json4s._
 import org.json4s.jackson.JsonMethods._