+++ src/test/scala/edu/rit/cs/dataunitylab/jsonoid/discovery/transformers/EnumTransformerSpec.scala
@@ -1,4 +1,5 @@
 package edu.rit.cs.dataunitylab.jsonoid.discovery
+package transformers
 
 import org.json4s.{DefaultFormats, Formats}